Finish up barebones template rendering
Due to lack of good documentation on how golang templates are
executed and/ord built it seems to be pretty hard to implement
layouts. The problem seems to boil down to the fact that its hard to
describe what is the base template that should be rendered as they are
just bunch of building blocks that are defined. So if there is a
template with nested component that is expected to be filled in by more
specific component (e.g. base html layout and body element) the template
has to first parse all wanted templates and then execute the most top
template. Bu the templating engine itself can not know which is the most
top one and has to be provided manually when multiple templates exist.

Basically solved this by adding a base "entrypoint" template that
defines general structure by using other components, so this way each
page template has to define the entrypoint which then will likely
execute layout template. Overall, pretty confusing and lack of good
documentation on templating doesn't help (or maybe I couldn't find any).