Add Google Drive support for backup storage
This adds ability to connect Google Drive account where the application
can store json backups. For this to work user has to create a google
drive api application with desktop application credentials and provide
DRIVE_ID and DRIVE_SECRET env variables with appropriate google
credentials.

Added basic migration support for database, so that old database
can be used with new version of the application which now has to store
drive refresh token.

Changed so that both post backup actions (json, gdrive) now have to be
enabled. These options are not exposed to the web ui, since for these
actions to work in the first place non changeable configuration has to
be provided, so it makes little sense to make them dynamically switchable.