Support saving Youtube playlists
Add a feature to also provide an ability to backup specified youtube
playlists. It works generally the same as Spotify just some fields
are different (e.g. no artist). Config wise, only way is providing
a playlist id.

While it does work, some parts are a bit rough. This is mainly
due to the fact that originally I only planned for a single source of
information, but now there are two places (Spotify, Youtube). To
simplify the implementation Youtube part is sort of an addon to the
Spotify part and is kinda attached with a band-aid.